requirements.txt
pandas==0.25.3
matplotlib==3.1.2
numpy==1.16.4
opencv_python==4.4.0.42
seaborn==0.10.0
tensorflow==2.3.0
Keras==2.4.3
scipy==1.4.1
tqdm==4.42.1
Pillow==8.2.0
scikit_learn==0.24.1
